Reorganize components
The main "components/" folder should only contain generic components that can be re-usable. Otherwise components scoped to a resource should be in the "pages/<resource>/components/" folder.
