+++ web/frontend/src/components/__tests__/Election.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-import ElectionForm from '../ElectionForm';
+import ElectionForm from '../../pages/election/components/ElectionForm';
 Enzyme.configure({ adapter: new Adapter() });
 
 let mockData = {
