+++ web/frontend/src/pages/ballot/Index.tsx
@@ -1,9 +1,9 @@
 import React, { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import SimpleTable from '../components/utils/SimpleTable';
-import { OPEN } from '../components/utils/StatusNumber';
-import './BallotIndex.css';
+import SimpleTable from '../../components/utils/SimpleTable';
+import { OPEN } from '../../components/utils/StatusNumber';
+import './Index.css';
 
 const BallotIndex: FC = () => {
   const { t } = useTranslation();
