+++ web/frontend/src/pages/ballot/Show.css
