+++ web/frontend/src/pages/ballot/Show.tsx
@@ -5,14 +5,14 @@ import kyber from '@dedis/kyber';
 import PropTypes from 'prop-types';
 import { Buffer } from 'buffer';
 
-import { ROUTE_BALLOT_INDEX } from '../Routes';
-import useElection from '../components/utils/useElection';
-import usePostCall from '../components/utils/usePostCall';
-import { ENDPOINT_EVOTING_CAST_BALLOT } from '../components/utils/Endpoints';
-import Modal from '../components/modal/Modal';
-import { OPEN } from '../components/utils/StatusNumber';
-import { encryptVote } from '../components/VoteEncrypt';
-import './BallotShow.css';
+import { ROUTE_BALLOT_INDEX } from '../../Routes';
+import useElection from '../../components/utils/useElection';
+import usePostCall from '../../components/utils/usePostCall';
+import { ENDPOINT_EVOTING_CAST_BALLOT } from '../../components/utils/Endpoints';
+import Modal from '../../components/modal/Modal';
+import { OPEN } from '../../components/utils/StatusNumber';
+import { encryptVote } from './components/VoteEncrypt';
+import './Show.css';
 
 const Ballot: FC = () => {
   var bytes: number[];
