+++ web/frontend/src/pages/ballot/components/VoteEncrypt.tsx
