+++ web/frontend/src/pages/election/Index.css
