+++ web/frontend/src/pages/election/New.tsx
@@ -1,8 +1,8 @@
 import React, { FC, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import ElectionForm from '../components/ElectionForm';
-import Modal from '../components/modal/Modal';
+import ElectionForm from './components/ElectionForm';
+import Modal from '../../components/modal/Modal';
 
 const ElectionCreate: FC = () => {
   const { t } = useTranslation();
