+++ web/frontend/src/pages/election/Show.css
