+++ web/frontend/src/pages/election/components/ElectionTable.tsx
@@ -14,8 +14,8 @@ import { useTranslation } from 'react-i18next';
 
 import Action from './Action';
 import Status from './Status';
-import ElectionFields from './utils/ElectionFields';
-import { ROUTE_ELECTION_SHOW } from '../Routes';
+import ElectionFields from '../../../components/utils/ElectionFields';
+import { ROUTE_ELECTION_SHOW } from '../../../Routes';
 
 type ElectionTableProps = {
   elections: string[];
