+++ web/frontend/src/pages/election/components/Status.css
