+++ web/frontend/src/pages/election/components/Status.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import useChangeStatus from './utils/useChangeStatus';
+import useChangeStatus from '../../../components/utils/useChangeStatus';
 import './Status.css';
 
 /*StatusSuccess is a class that acts as a container for the display of the
