+++ web/frontend/src/pages/result/Show.css
