+++ web/frontend/src/pages/result/components/Result.tsx
